move a lot of shared functionality to the base class
