+++ src/libs/loco/include/loco/controller/KinematicTrackingController.h
@@ -26,82 +26,53 @@ public:
 
     gui::SimpleGroundModel ground{};
 
-    std::shared_ptr<LeggedRobot> robot;
-    std::shared_ptr<IK_Solver> ikSolver = nullptr;
-    GeneralizedCoordinatesRobotRepresentation gcrr;
-
     //Vector of point/radius/color
     std::vector<std::tuple<P3D, double, V3D>> drawList{};
-
-    double heelHeight;
-    double toeHeight;
-
-    const double pi = 3.14159265358979323846;
-    const double twoPi = pi * 2.0;
-        
-    double t = 0;
-    double cycleLength = 0.8; // in seconds
-
-    double stanceStart = 0.0; // in percent
-    double swingStart = 0.5; // in percent
-    double heelStrikeStart = 0.9; // in percent
-
-    double heelToeDistance;
 
     std::array<P3D, 2> stanceTargets;
     std::array<P3D, 2> swingTargets;
     std::array<P3D, 2> heelStrikeTargets;
     std::array<Trajectory3D, 2> swingTrajectories;
-    std::array<std::array<int, 5>, 2> footJointIndices;
+
     std::array<P3D, 2> toeTargets;
     std::array<P3D, 2> toeStrikeTarget;
     std::array<P3D, 2> heelTargets;
     std::array<P3D, 2> heelStarts;
     std::array<P3D, 2> heelEnds;
-//    std::array<double, 2> heelHeight;
+
     std::array<bool, 2> heelStartSet;
 
-    double shoulderAmplitudeDegree = 5.0; // in degree
-    double shoulderAmplitude = shoulderAmplitudeDegree * twoPi / 360.0;
-    double shoulderMin = shoulderAmplitude;
-    double shoulderMax = -shoulderAmplitude; // negative rotation: moves arm forward
     Trajectory1D shoulderJointTrajectory;
-
-    double elbowMin = -0.25;
-    double elbowMax = -0.45;
-    double elbowswingoffset = 0.03;
     Trajectory1D elbowJointTrajectory;
 
-    double dt = 1./30;
-
-    // 1: shoulder sagittal plane, 2: elbow saggital plane, 3: shoulder torsion
-    std::array<std::array<int, 3>, 2> armJointIndices;
-
-    // for upper body
-    int spineIdx;
-    double spineAmplitudeDegree = 2.0;
-    double spineAmplitude = toRad(spineAmplitudeDegree);
-    int neckIdx;
-
-    enum Phase {Stance, Swing, HeelStrike};
     int its = 0;
 
 public:
     /**
      * constructor
      */
-    explicit KinematicTrackingController(const std::shared_ptr<LocomotionTrajectoryPlanner> &planner) : LocomotionController(planner), gcrr(planner->robot) {
-        this->robot = planner->robot;
-        ikSolver = std::make_shared<IK_Solver>(robot);
-
-        auto toe = robot->getLimb(0);
-        auto heel = robot->getLimb(2);
-
-        heelToeDistance = V3D(heel->getEEWorldPos(), toe->getEEWorldPos()).norm();
-
-        heelHeight = (heel->limbRoot->getWorldCoordinates() + heel->defaultEEOffsetWorld)[1];
-        toeHeight = (toe->limbRoot->getWorldCoordinates() + toe->defaultEEOffsetWorld)[1];
-
+    explicit KinematicTrackingController(
+        const std::shared_ptr<LocomotionTrajectoryPlanner> &planner,
+        double cycleLength,
+        double stanceStart,
+        double swingStart,
+        double heelStrikeStart,
+        double shoulderAmplitudeDegree = 5.0,
+        double elbowMin = -0.25,
+        double elbowMax = -0.45,
+        double spineAmplitudeDegree = 2.0
+    ) 
+        : LocomotionController(
+            planner,
+            cycleLength,
+            stanceStart,
+            swingStart,
+            heelStrikeStart,
+            shoulderAmplitudeDegree,
+            elbowMin,
+            elbowMax,
+            spineAmplitudeDegree
+        ) {
         for (int i : {0, 1}) {
             P3D pToes = robot->getLimb(i)->getEEWorldPos();
             P3D pHeel = robot->getLimb(i + 2)->getEEWorldPos();
@@ -118,42 +89,13 @@ public:
             swingTrajectories[i].addKnot(1, V3D(pHeel));
         }
 
-        std::array lLegJoints = {"lHip_1", "lKnee", "lAnkle_1", "lToeJoint", "lAnkle_2"};
-        std::array rLegJoints = {"rHip_1", "rKnee", "rAnkle_1", "rToeJoint", "rAnkle_2"};
-        for (int i = 0; i < lLegJoints.size(); ++i) {
-            int lIdx = robot->getJointIndex(lLegJoints[i]);
-            int rIdx = robot->getJointIndex(rLegJoints[i]);
-            footJointIndices[0][i] = rIdx;
-            footJointIndices[1][i] = lIdx;
-        }
-
-
-
-        spineIdx = robot->getJointIndex("upperback_y");
-        neckIdx = robot->getJointIndex("lowerneck_y");
-
-        std::array lArmJoints = {"lShoulder_1", "lElbow_flexion_extension", "lShoulder_torsion"};
-        std::array rArmJoints = {"rShoulder_1", "rElbow_flexion_extension", "rShoulder_torsion"};
-        for (int i = 0; i < lArmJoints.size(); ++i) {
-            int lIdx = robot->getJointIndex(lArmJoints[i]);
-            int rIdx = robot->getJointIndex(rArmJoints[i]);
-            armJointIndices[0][i] = rIdx;
-            armJointIndices[1][i] = lIdx;
-        }
-
-        // negative degrees raise the arms
-        // max elongation should roughly be in sync with the furthest position of the knee (at ~65% of the cycle); maybe not?
-        //shoulderJointTrajectory.addKnot(stanceStart, shoulderMax);
-        //shoulderJointTrajectory.addKnot(0.65, shoulderMax);
-        //shoulderJointTrajectory.addKnot(1.0, 0.0);
-
         shoulderJointTrajectory.addKnot(stanceStart, shoulderMax);
         shoulderJointTrajectory.addKnot(swingStart, shoulderMin);
         shoulderJointTrajectory.addKnot(heelStrikeStart, shoulderMax);
 
         elbowJointTrajectory.addKnot(stanceStart, elbowMax);
-        elbowJointTrajectory.addKnot(swingStart + elbowswingoffset, elbowMin);
-        elbowJointTrajectory.addKnot(heelStrikeStart + elbowswingoffset, elbowMax);
+        elbowJointTrajectory.addKnot(swingStart + elbowSwingOffset, elbowMin);
+        elbowJointTrajectory.addKnot(heelStrikeStart + elbowSwingOffset, elbowMax);
     }
 
     /**
@@ -161,9 +103,6 @@ public:
      */
     ~KinematicTrackingController() override = default;
 
-    double toRad(double degree) const {
-        return degree * twoPi / 360.0;
-    }
     void generateMotionTrajectories(double dt = 1.0 / 30) override {
         planner->planGenerationTime = planner->simTime;
         planner->generateTrajectoriesFromCurrentState(dt);
@@ -171,25 +110,15 @@ public:
 
 
     void computeAndApplyControlSignals(double deltaT) override {
-        // set base pose. in this assignment, we just assume the base perfectly
-        // follow target base trajectory.
-
-        dt = deltaT;
-
+        double dt = deltaT;
         gcrr.syncGeneralizedCoordinatesWithRobotState();
         P3D targetPos = planner->getTargetTrunkPositionAtTime(planner->getSimTime() + dt);
         Quaternion targetOrientation = planner->getTargetTrunkOrientationAtTime(planner->getSimTime() + dt);
         robot->setRootState(targetPos + V3D(0, ground.getHeight(targetPos), 0), targetOrientation);
 
-        
         gcrr.syncGeneralizedCoordinatesWithRobotState();
         t += dt;
 
-
-        // use the toe as the targeted endeffector in the stance phase
-        // use the heel as the targeted endeffector in the swing phase
-        // idx i == 0 <-> right leg
-        // before the phase
         for (int i : {0, 1}) {
             Phase currentPhase = getPhase(i, t);
 
@@ -221,7 +150,6 @@ public:
                 setAnkleAngleDuringHeelStrike(i);
             }
         }
-
         // preparation for the next phase depending on the transition
         for (int i : {0, 1}) {
             Phase currentPhase = getPhase(i, t);
@@ -235,10 +163,6 @@ public:
                 toeTargets[i] = computeToeTarget(i, nextPhase);
                 heelTargets[i] = computeHeelTarget(i, nextPhase);
             }
-            /*
-            toeTargets[i] = computeToeTarget(i, nextPhase);
-            heelTargets[i] = computeHeelTarget(i, nextPhase);
-            */
             if(nextPhase == Phase::Stance && isEarlyStance(getCyclePercent(i, t+dt)))
                 computeEarlyStanceHeelStrike(i);
         }
@@ -295,7 +219,6 @@ public:
         gcrr.syncRobotStateWithGeneralizedCoordinates();
 
     }
-
 
     void initializeSwingPhase(int i) {
         V3D velocity = planner->getTargetTrunkVelocityAtTime(planner->getSimTime() /* + dt*/);
@@ -355,16 +278,6 @@ public:
         double cyclePercent = getCyclePercent(i, t);
         // return lerp(heelStarts[i], heelEnds[i], remap(cyclePercent, swingStart, heelStrikeStart));
         return getP3D(swingTrajectories[i].evaluate_catmull_rom(remap(cyclePercent, swingStart, heelStrikeStart)));
-    }
-
-    void setHeelTarget(int i, P3D pTarget) {
-        auto heel = robot->getLimb(i + 2);
-        ikSolver->addEndEffectorTarget(heel->eeRB, heel->ee->endEffectorOffset, pTarget);
-    }
-    
-    void setToeTarget(int i, P3D pTarget) {
-        auto foot = robot->getLimb(i);
-        ikSolver->addEndEffectorTarget(foot->eeRB, foot->ee->endEffectorOffset, pTarget); 
     }
 
     P3D computeToeTarget(int i, Phase nextPhase) {
@@ -503,107 +416,11 @@ public:
         gcrr.syncRobotStateWithGeneralizedCoordinates();
     }
 
-    void computeStanceTarget(int i) {
-        auto foot = robot->getLimb(i);
-        stanceTargets[i] = foot->getEEWorldPos();
-    }
-
-    void setStanceTarget(int i) {
-        auto foot = robot->getLimb(i);
-        ikSolver->addEndEffectorTarget(foot->eeRB, foot->ee->endEffectorOffset, stanceTargets[i]);
-    }
-    
-    void computeSwingTrajectory(int i) {
-        auto heel = robot->getLimb(i + 2);
-
-        swingTrajectories[i].clear();
-        P3D pStart = heel->getEEWorldPos();
-        V3D offset(pStart, heel->limbRoot->getWorldCoordinates() + heel->defaultEEOffsetWorld);
-        P3D pEnd = pStart + offset * 2.1;
-        swingTrajectories[i].addKnot(0, V3D(pStart));
-        swingTrajectories[i].addKnot(1, V3D(pEnd));
-    }
-
-    void computeSwingTarget(int i) {
-        auto heel = robot->getLimb(i + 2);
-        P3D pStart = heel->getEEWorldPos();
-        P3D pEnd = getP3D(robot->getHeading() * V3D(heel->limbRoot->getWorldCoordinates() + heel->defaultEEOffsetWorld  // heel position in default pose
-                    + 0.05 * robot->getForward()));             // offset in heading direction
-
-
-        swingTargets[i] = lerp(pStart, pEnd, remap(getCyclePercent(i, t), swingStart, heelStrikeStart));
-    }
-
-    void lockSwingTargetForHeelStrike(int i) {
-        auto heel = robot->getLimb(i + 2);
-        P3D pHeel = heel->getEEWorldPos();
-        swingTargets[i] = pHeel;
-    }
-
-    void setSwingTarget(int i) {
-        auto heel = robot->getLimb(i + 2);
-        ikSolver->addEndEffectorTarget(heel->eeRB, heel->ee->endEffectorOffset, swingTargets[i]);
-    }
-
     void computeEarlyStanceHeelStrike(int i) {
         auto toes = robot->getLimb(i);
         auto heel = robot->getLimb(i + 2);
         heelTargets[i] = heel->getEEWorldPos();
         heelTargets[i][1] = heelHeight + ground.getHeight(heelTargets[i]);
-    }
-
-    void computeHeelStrikeTarget(int i) {
-        auto heel = robot->getLimb(i + 2);
-        heelStrikeTargets[i] = heel->getEEWorldPos() + robot->getForward() * heelToeDistance;
-    }
-
-    void setHeelStrikeTarget(int i) {
-        auto toes = robot->getLimb(i);
-        double cyclePercent = getCyclePercent(i, t);
-        // maybe fix the starting position?
-        P3D pTarget = lerp(toes->getEEWorldPos(), heelStrikeTargets[i], remap(cyclePercent, heelStrikeStart, 1.0));
-        ikSolver->addEndEffectorTarget(toes->eeRB, toes->ee->endEffectorOffset, pTarget);
-    }
-
-    // remap t in [a, b] to [0, 1]
-    static double remap(double time, double a, double b) {
-        return (time - a) / (b - a);
-    }
-
-    template <typename T>
-    T lerp(T  a, T b, double time) {
-        return a * (1.0 - time) + b * time;
-    }
-
-    double getCyclePercent(int i, double time) const {
-        return std::fmod(i * 0.5 * cycleLength + time, cycleLength) / cycleLength;
-    }
-
-    bool isStance(double cyclePercent) const {
-        return stanceStart <= cyclePercent && cyclePercent < swingStart;
-    }
-
-    bool isEarlyStance(double cyclePercent) const {
-        return isStance(cyclePercent) && (cyclePercent < (stanceStart + swingStart)*0.5);
-    }
-
-    bool isSwing(double cyclePercent) const {
-        return swingStart <= cyclePercent && cyclePercent < heelStrikeStart;
-    }
-    
-    bool isHeelStrike(double cyclePercent) const {
-        return heelStrikeStart <= cyclePercent && cyclePercent < 1.0;
-    }
-
-    Phase getPhase(int i, double time) const {
-        double cyclePercent = getCyclePercent(i, time);
-        if (isStance(cyclePercent)) {
-            return Phase::Stance;
-        } else if (isSwing(cyclePercent)) {
-            return Phase::Swing;
-        } else if (isHeelStrike(cyclePercent)) {
-            return Phase::HeelStrike;
-        }
     }
 
     void setToesToFloor(int i, const P3D& position) {
@@ -615,7 +432,7 @@ public:
     }
 
     void advanceInTime(double deltaT) override {
-        dt = deltaT;
+        double dt = deltaT;
         planner->advanceInTime(dt);
     }
