+++ src/libs/loco/include/loco/controller/LocomotionController.h
@@ -2,6 +2,7 @@
 
 #include <crl-basic/gui/renderer.h>
 #include <loco/planner/LocomotionTrajectoryPlanner.h>
+#include <loco/kinematics/IK_Solver.h>
 
 namespace crl::loco {
 
@@ -12,17 +13,99 @@ namespace crl::loco {
  */
 class LocomotionController {
 public:
-    std::shared_ptr<LocomotionTrajectoryPlanner> planner = nullptr;
+    const double pi = 3.14159265358979323846;
+    const double twoPi = pi * 2.0;
 
+    std::shared_ptr<LocomotionTrajectoryPlanner> planner = nullptr;
+    GeneralizedCoordinatesRobotRepresentation gcrr;
+    std::shared_ptr<LeggedRobot> robot;
+    std::shared_ptr<IK_Solver> ikSolver = nullptr;
+
+    double t;
+    
+    const double cycleLength;
+    const double stanceStart;
+    const double swingStart;
+    const double heelStrikeStart;
+    enum Phase { Stance, Swing, HeelStrike };
+
+    // 1: shoulder sagittal plane, 2: elbow saggital plane, 3: shoulder torsion
+    std::array<std::array<int, 3>, 2> armJointIndices;
+    const double shoulderAmplitude;
+    const double shoulderMin;
+    const double shoulderMax;
+
+    const double elbowMin;
+    const double elbowMax;
+    const double elbowSwingOffset = 0.03;
+
+    const double spineAmplitude;
+    int spineIdx;
+    int neckIdx;
+
+    // hip, knee, ankle1, toe, ankle2
+    std::array<std::array<int, 5>, 2> footJointIndices;
+    double heelHeight;
+    double toeHeight;
+    double heelToeDistance;
 public:
-    LocomotionController(const std::shared_ptr<LocomotionTrajectoryPlanner>& planner) : planner(planner) {}
+    LocomotionController(const std::shared_ptr<LocomotionTrajectoryPlanner>& planner, double cycleLength, double stanceStart, double swingStart,
+                         double heelStrikeStart, double shoulderAmplitudeDegree, double elbowMin, double elbowMax, double spinneAmplitudeDegree)
+        : planner(planner),
+          gcrr(planner->robot),
+          cycleLength(cycleLength),
+          stanceStart(stanceStart),
+          swingStart(swingStart),
+          heelStrikeStart(heelStrikeStart),
+          shoulderAmplitude(toRad(shoulderAmplitudeDegree)),
+          shoulderMin(toRad(shoulderAmplitudeDegree)),
+          shoulderMax(-toRad(shoulderAmplitudeDegree)),
+          elbowMin(elbowMin),
+          elbowMax(elbowMax),
+          spineAmplitude(toRad(spinneAmplitudeDegree))
+    {
+        
+        t = 0.0;
+
+        robot = planner->robot;
+        ikSolver = std::make_shared<IK_Solver>(robot);
+
+        auto toe = robot->getLimb(0);
+        auto heel = robot->getLimb(2);
+
+        heelToeDistance = V3D(heel->getEEWorldPos(), toe->getEEWorldPos()).norm();
+        heelHeight = (heel->limbRoot->getWorldCoordinates() + heel->defaultEEOffsetWorld)[1];
+        toeHeight = (toe->limbRoot->getWorldCoordinates() + toe->defaultEEOffsetWorld)[1];
+
+        std::array lLegJoints = {"lHip_1", "lKnee", "lAnkle_1", "lToeJoint", "lAnkle_2"};
+        std::array rLegJoints = {"rHip_1", "rKnee", "rAnkle_1", "rToeJoint", "rAnkle_2"};
+        for (int i = 0; i < lLegJoints.size(); ++i) {
+            int lIdx = robot->getJointIndex(lLegJoints[i]);
+            int rIdx = robot->getJointIndex(rLegJoints[i]);
+            footJointIndices[0][i] = rIdx;
+            footJointIndices[1][i] = lIdx;
+        }
+
+        std::array lArmJoints = {"lShoulder_1", "lElbow_flexion_extension", "lShoulder_torsion"};
+        std::array rArmJoints = {"rShoulder_1", "rElbow_flexion_extension", "rShoulder_torsion"};
+        for (int i = 0; i < lArmJoints.size(); ++i) {
+            int lIdx = robot->getJointIndex(lArmJoints[i]);
+            int rIdx = robot->getJointIndex(rArmJoints[i]);
+            armJointIndices[0][i] = rIdx;
+            armJointIndices[1][i] = lIdx;
+        }
+
+        spineIdx = robot->getJointIndex("upperback_y");
+        neckIdx = robot->getJointIndex("lowerneck_y");
+
+    }
 
     virtual ~LocomotionController(){};
 
     /**
      * Generate motion trajectory with timestep size dt.
      */
-    virtual void generateMotionTrajectories(double dt) = 0;
+    virtual void generateMotionTrajectories(double dt = 1.0 / 30.0) = 0;
 
     /**
      * Compute and apply control signal with timestep size dt.
@@ -52,6 +135,62 @@ public:
      * e.g. joint torque etc.
      */
     virtual void plotDebugInfo() = 0;
+
+
+    double toRad(double degree) const {
+        return degree * twoPi / 360.0;
+    }
+    
+    // remap t in [a, b] to [0, 1]
+    static double remap(double time, double a, double b) {
+        return (time - a) / (b - a);
+    }
+
+    template <typename T>
+    T lerp(T a, T b, double time) {
+        return a * (1.0 - time) + b * time;
+    }
+
+    double getCyclePercent(int i, double time) const {
+        return std::fmod(i * 0.5 * cycleLength + time, cycleLength) / cycleLength;
+    }
+
+    bool isStance(double cyclePercent) const {
+        return stanceStart <= cyclePercent && cyclePercent < swingStart;
+    }
+
+    bool isEarlyStance(double cyclePercent) const {
+        return isStance(cyclePercent) && (cyclePercent < (stanceStart + swingStart) * 0.5);
+    }
+
+    bool isSwing(double cyclePercent) const {
+        return swingStart <= cyclePercent && cyclePercent < heelStrikeStart;
+    }
+
+    bool isHeelStrike(double cyclePercent) const {
+        return heelStrikeStart <= cyclePercent && cyclePercent < 1.0;
+    }
+
+    Phase getPhase(int i, double time) const {
+        double cyclePercent = getCyclePercent(i, time);
+        if (isStance(cyclePercent)) {
+            return Phase::Stance;
+        } else if (isSwing(cyclePercent)) {
+            return Phase::Swing;
+        } else if (isHeelStrike(cyclePercent)) {
+            return Phase::HeelStrike;
+        }
+    }
+
+    void setHeelTarget(int i, P3D pTarget) {
+        auto heel = robot->getLimb(i + 2);
+        ikSolver->addEndEffectorTarget(heel->eeRB, heel->ee->endEffectorOffset, pTarget);
+    }
+
+    void setToeTarget(int i, P3D pTarget) {
+        auto foot = robot->getLimb(i);
+        ikSolver->addEndEffectorTarget(foot->eeRB, foot->ee->endEffectorOffset, pTarget);
+    }
 };
 
 }  // namespace crl::loco
